requirements.txt
appdirs==1.4.3
backcall==0.1.0
-e git+https://github.com/kivy/buildozer.git@d3628f1568a4a3c50465ec817a6d82c52817a3bb#egg=buildozer
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
Cython==0.28.1
decorator==4.2.1
docutils==0.14
idna==2.6
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.11.1
Jinja2==2.10
jnius==1.1.0
Kivy==1.10.0
Kivy-Garden==0.1.4
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
numpy==1.14.2
opencv-python==3.4.0.12
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.1.0
plyer==1.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pygame==1.9.3
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.2
python-for-android==0.6.0
python-interface==1.4.0
pytz==2018.4
readline==6.2.4.1
requests==2.18.4
sh==1.12.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
virtualenv==15.2.0
wcwidth==0.1.7
zope.event==4.3.0
zope.interface==4.4.3
zope.schema==4.5.0
